removed DevicePool completely moving it's logic into IDriverPool interface
