+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/gui/mobile/devices/android/phone/pages/notifications/NotificationPage.java
@@ -29,8 +29,8 @@ import com.qaprosoft.carina.core.foundation.utils.android.AndroidService;
 import com.qaprosoft.carina.core.foundation.utils.factory.DeviceType;
 import com.qaprosoft.carina.core.foundation.utils.mobile.MobileUtils;
 import com.qaprosoft.carina.core.foundation.utils.mobile.notifications.android.Notification;
+import com.qaprosoft.carina.core.foundation.webdriver.IDriverPool;
 import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
-import com.qaprosoft.carina.core.foundation.webdriver.device.DevicePool;
 import com.qaprosoft.carina.core.gui.mobile.devices.MobileAbstractPage;
 
 import io.appium.java_client.MobileBy;
@@ -154,7 +154,7 @@ public class NotificationPage extends MobileAbstractPage {
     public String getItemText(int num) {
         try {
             LOGGER.info("Visible text:" + lastItemsContent.get(num).findExtendedWebElements(MobileBy.className("android.widget.TextView")).size());
-            if (DevicePool.getDevice().getDeviceType() == DeviceType.Type.ANDROID_TABLET) {
+            if (IDriverPool.getDevice().getDeviceType() == DeviceType.Type.ANDROID_TABLET) {
                 try {
                     if (lastItemsContent.get(num).findExtendedWebElement(MobileBy.id(itemText_Tablet_Locator_Text)).isElementNotPresent(1)) {
                         return lastItemsContent.get(num).findExtendedWebElement(MobileBy.id(itemText_Phone_Locator_Text)).getText();
