+++ carina-webdriver/src/test/java/com/qaprosoft/carina/core/foundation/webdriver/DriverPoolTest.java
@@ -30,7 +30,6 @@ import com.qaprosoft.carina.core.foundation.utils.Configuration;
 import com.qaprosoft.carina.core.foundation.utils.Configuration.Parameter;
 import com.qaprosoft.carina.core.foundation.utils.R;
 import com.qaprosoft.carina.core.foundation.webdriver.TestPhase.Phase;
-import com.qaprosoft.carina.core.foundation.webdriver.device.DevicePool;
 
 public class DriverPoolTest implements IDriverPool {
 
@@ -239,7 +238,7 @@ public class DriverPoolTest implements IDriverPool {
 
         // new 6.0 approach to manipulate drivers via regular Set
         //TODO: cover by unit tests device pool registration for each driver
-        CarinaDriver carinaDriver = new CarinaDriver(name, driver, DevicePool.getNullDevice(), TestPhase.getActivePhase(), threadId);
+        CarinaDriver carinaDriver = new CarinaDriver(name, driver, IDriverPool.getNullDevice(), TestPhase.getActivePhase(), threadId);
         driversPool.add(carinaDriver);
     }
